requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pymongo==4.13.2
python-dotenv==1.1.1
email-validator==2.2.0
httpx==0.28.1
